data_pipeline/requirements.txt
requests==2.32.3
azure-identity==1.19.0
azure-ai-ml==1.22.4
python-dotenv==1.0.1
pyodbc==5.2.0
azure-keyvault==4.2.0
ruff==0.8.0